+++ static/etc/init.d/hhvm.sh
@@ -29,8 +29,8 @@ SCRIPTNAME=/etc/init.d/$NAME
 # Default values. This values can be overwritten in '/etc/default/hhvm'
 CONFIG_FILE="/etc/hhvm/server.ini"
 SYSTEM_CONFIG_FILE="/etc/hhvm/php.ini"
-RUN_AS_USER="hipstack"
-RUN_AS_GROUP="hipstack"
+RUN_AS_USER="hhvm"
+RUN_AS_GROUP="hhvm"
 ADDITIONAL_ARGS=""
 
 # Read configuration variable file if it is present