+++ changes.md
@@ -1,4 +1,5 @@
 #### 0.2.0
+* Tweaks to apache2.conf.
 * Added PHP prepend and append files.
 * Bundled Memcached and added Memcache(d) benchmark tests. (@source http://www.leaseweblabs.com/2013/03/memcache-vs-memcached-php-benchmark/)
 * Added NewRelic source.